requirements.txt
flask
python-dotenv
langchain
google-generativeai
PyPDF2
